refactor: popup refactored to RootProvider and Router, useContext for SelectedCountries
